analizer v1.

+++ main/src/analizer.cpp
@@ -1,7 +1,7 @@
 #include "analizer.h"
 
 
-    ana::Analizer::Analizer(uint8_t CoreNum)
+    ana::Analizer::Analizer(int CoreNum)
 {
 
 }
